refactored search from 'god file' into different processes other tools may want to use
