+++ gui/search/__init__.py
@@ -1,0 +1,3 @@
+from gui.search._window import SearchWindow
+
+__all__ = ["SearchWindow"]
